requirements.txt
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
decorator==5.1.1
Django==3.2.18
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
psycopg2-binary==2.9.6
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sqlparse==0.4.3
tornado==6.3.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.13
